sidepanel_schema
